labelme/requirements.txt
numpy==1.21.1
opencv-python-headless==4.5.3.56
Pillow==8.3.1
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyYAML==5.4.1
QtPy==1.9.0
termcolor==1.1.0
